Use a sorted list for the variable constant values
Something about javac or ant doesn't cope very well if static
constants in a class are changed so we want to keep them as consistent
as possible. This changes it so the index values are calculated
upfront based on the sorted order of the names before generating the
trie so that they won't change unless a variable is added or removed.