Add a singleton class to load articles
The loader maintains a cache of the last 8 articles so that it won't
go crazy repeatedly loading articles while it tries to expand variable
substitutions for parent articles. The article activity now has the
basics to use the article loader and sets the view title to the
infinitive of the article's verb.