bash-snippets: Support two params in remove_group and generate_group
If an extra argument is given to remove_group or generate_group it
will now try to match groups that have a second parameter which
contains part of the beginning of the first parameter. It seems that
these are often used in Viccionari to specify different stems that are
only different in Balearic Catalan.